[CHORE] view 분리 (#441)
